+++ phpunit/bootstrap.php
@@ -86,13 +86,14 @@ tests_add_filter( 'wp_die_handler', 'fail_if_died' );
 
 $GLOBALS['wp_tests_options'] = array(
 	'gutenberg-experiments' => array(
-		'gutenberg-widget-experiments' => '1',
-		'gutenberg-full-site-editing'  => 1,
-		'gutenberg-form-blocks'        => 1,
-		'gutenberg-block-experiments'  => 1,
-		'gutenberg-media-processing'   => 1,
-		'gutenberg-guidelines'         => 1,
-		'gutenberg-dashboard-widgets'  => 1,
+		'gutenberg-widget-experiments'      => '1',
+		'gutenberg-full-site-editing'       => 1,
+		'gutenberg-form-blocks'             => 1,
+		'gutenberg-block-experiments'       => 1,
+		'gutenberg-media-processing'        => 1,
+		'gutenberg-guidelines'              => 1,
+		'gutenberg-dashboard-widgets'       => 1,
+		'gutenberg-real-time-collaboration' => 1,
 	),
 );